[refactor] 파일 구조 변경 및 불필요한 코드 삭제
