+++ backend/src/route/app.ts
@@ -2,8 +2,7 @@ import express from 'express';
 import cookieParser from 'cookie-parser';
 import passport from 'passport';
 import morgan from 'morgan';
-import loginRouter from './user/login';
-import userRouter from './user/user';
+import userRouter from './user';
 import mainRouter from './main';
 import schoolRouter from './school';
 
@@ -14,7 +13,6 @@ app.use(express.json());
 app.use(passport.initialize());
 app.use(cookieParser());
 
-app.use('/api/login', loginRouter);
 app.use('/api/user', userRouter);
 app.use('/api/main', mainRouter);
 app.use('/api/school', schoolRouter);
